Add the DPO indicator

+++ src/main/java/ru/ilysenko/tinka/indicator/MomentumIndicator.java
@@ -46,12 +46,12 @@ public class MomentumIndicator extends AbstractIndicator {
     }
 
     @Override
-    public int getOverboughtThreshold() {
+    public double getOverboughtThreshold() {
         return OVERBOUGHT_THRESHOLD;
     }
 
     @Override
-    public int getOversoldThreshold() {
+    public double getOversoldThreshold() {
         return OVERSOLD_THRESHOLD;
     }
 }
